refactor: reduce the number of page components
